Add admin login page.
The login page shows input texts to enter username and password. It also shows
"Remember me" checkbox to save the username so that a user does not need to
input the username again. "Remember me" button does not work at the moment.

Login action stores the username in the session so that subsequent actions can
access it easily. Currently, there is no password checking. This should be
added once we have a database backend. Also there are lots of FIXMEs to be
done.