Add a method to Global to return a mongodb connection.
Override GlobalSettings.onStart to create a MongoConnection and add a method to
return the instance.

MongoDriver and MongoConnection involve creation costs the driver may create a
new ActorSystem, and the connection will connect to the servers. It is a good
idea to store the driver and the connection to reuse them.