Change the supervision strategy of UserActionActor router.
Routers default to a strategy of "always escalate". This is problematic because
a failure in a routee is escalated up to the router's supervisor for handling.
If the router's supervisor decides to restart the child, (which is the default,
unfortunately), the router and all of its routees are restarted.

So override the router's strategy with SupervisorStrategy.defaultStrategy which
restarts only the failing child actor upon Exception.

See Routers and Supervision section of
http://doc.akka.io/docs/akka/2.2.1/scala/routing.html for further discussions.