Add RoutingTableView.
RoutingTableView will be used to determine whether the request will be
handled in this server or not.
Each server has hashing value and handles range in (hash of previous
server, hash of current server]. If requested hash is in range,
queryServerToHandle method returns HandleHere. Otherwise returns
HandleIn(address).

And this patch also extracts methods, types and a class from
RoutingTableBuilder to package object. Because they are used in both
RoutingTableBuilder and RoutingTableView.