Add RoutingTableView.
RoutingTableView will be used to determine whether a request will be handled in
the current server. Each server has a hash value and handles reuqests ranged in
(hash of the previous server, hash of the current server]. If the hash of a
request is in the given range, queryServerToHandle method returns HandleHere.
Otherwise it returns HandleIn(address).

This patch extracts methods, types and a class from RoutingTableBuilder to
package object because they are used in both RoutingTableBuilder and
RoutingTableView.