Implement RoutingTableWorker
RoutingTableWorker uses Curator to synchronize consistent hash routing table.
Worker creates a sequential ephemeral node at /workers with protection mode
when it starts. Protection mode is one of the curator recipe. It adds prefix
with client id to ensure that the node is created. A worker watches the
routing table node and updates RoutingTable when the routing table node is
changed.