Broadcasts UpdateRoutingTable messages to UserActionActor.
When RoutingTableWorker detect routing table is changed, it sends a
update message to parent(BeyondSupervisor). And BeyondSupervisor
broadcasts this message to all UserActionActor.

This patch includes two refactorings.
The one is moving case class declaration, that are sent to
UserActionActor, to companion object of UserActionActor.
And the other is extracting names of actors, that are used in
BeyondSupervisor.