Response redirect code when the request is not in consistent hashing range of the server.
If the hash of request is not in current server's hashing range, the
server have to response the address of server that will handle this
request.
If a client receives RedirectStatusCode(currently 310. I just pick one
of the redirection code not used in HTTP protocol.), The client have to
retry the address in response.

This patch also moves hashing method to RequestWithUsername from
BlockAndRequest. BlockAndRequest uses hashing method in
RequestWithUsername.