Use an empty array as default data of RoutingTable in ZooKeeper.
Currently, JsonParseException is often thrown. This is because I parsed data in
the RoutingTable node only if it exists. I expected znode has null data when
znode is created without any data passed in, but Curator passes the client IP
as default if no data is given.

This patch passes an empty JsArray as data when RoutingTalbe node is created.