Use empty array as default data of RoutingTable in zookeeper.
Currently, JsonParseException is often occured.
This is because I parsed data in the RoutingTable node if exists.
I expected znode has null data, when znode is created without passed
data. But Curator passes client IP as data if no data is passed.

This patch passes empty JsArray as data when RoutingTalbe node is
created.