Add an extractor for RouteAddress.
This patch introduces an extractor for RouteAddress which is handy for parsing
RouteAddress into IP and port. The extractor will be used later in management
console to gather list of workers.

Change the return type of BeyondConfiguration.currentServerAddress to
RouteAddress to make it explicit that it returns an instance of RouteAddress
that can be parsed with the extractor.