Make ZooKeeperLauncher fork ZooKeeperServer.
Currently, the lifecycle of ZooKeeper server is exactly the same to that of
Beyond instance. So whenever Beyond restarts, the ZooKeeper server also
restarts.

Unfortunately, this dependency causes unnecessary delay in Beyond shutdown.
When shutting down Beyond, the ZooKeeper server can be stopped before Apache
Curator connection to the ZooKeeper server is closed. In this case, Curator
tries to reconnect to the server for each pending operation, which causes
significant delay in shutdown.

This patch is the first step to separate the lifecycle of ZooKeeper server from
that of Beyond instance. ZooKeeperServerMainWithPIDFile is the main class to
execute ZooKeeper Server. It writes PID to the configured path and executes
ZooKeeper server. The next patch will make ZooKeeperLauncher kill the process
if it is not responding.