Add ProcessCommand.
ProcessCommand is introduced to provide OS dependent process commands. This
patch implements sending a terminal signal only because it is the only
command we need at this time.

ProcessCommand has a factory method to retrieve the OS specific ProcessCommand.
One can retrieve the instance using the apply method of ProcessCommand.