Add NumberOfRequestsPerSecond.
NumberOfRequestsPerSecond provides a counter which represents the number of
requests per second.

Internally, it has two counters for efficiency: front count and back count.
getNumberOfRequestsPerSecond method returns only the front count.
increase method does not increase the front count directly. Instead, it
increases the back count. Every second, the front count is updated with the
value of the back count and the back count is initialized back to 0.