Add TickGenerator.
TickGenerator sends a Tick message to itself with interval.
TickGenerator is self-typed trait for Actor, so it has to be mixed with
Actor.

TickGenerator adds hook on postStop of Actor, but self typed trait cannot
call the method of mixed trait. So TickGenerator extends PostStopCaller
that has only one abstract method postStop.

TickGenerator's postStop method is an abstract method, because
TickGenerator inherits PostStopCaller and it's postStop method is an
abstract method.
The class who mixes TickGenerator must implements postStop. And it is
not a problem, because TickGenerator is self-typed trait of Actor.