Add TestGameClient.
TestGameClient is a command line tool to test Beyond actions. After entering
username and password, we can request actions with action URL and parameters.

TestGameClient internally uses sprary-client to asynchronously send HTTP
requests. Currently, the base URL is hardcoded to http://localhost:9000, but
this will be changed to a command line argument in the follow-up patch.