Add console library.
console has five methods for logging: log, info, warn, debug and error.

In Rhino, these methods must be implemented with variable length arguments. But
this is not possible in Scala because Scala implements variable length
arguments with Seq which is not compatible with Rhino.

To circumvent this limitation, changed console to pass a string after
performing format conversion in JavaScript.