Share GlobalScope among GamePluginWorkers.
Currently, each GamePluginWorker has their own ModuleScope. It makes the
recovery of scope easy because simply restarting the worker recovers the scope.

However, we implemented the timer API and will implement the Future API in a
way that these APIs don't ensure the callback is executed in the same worker.
So we decided to share the scope among GamePluginWorkers.