Add some of DB API(insert, find and findOne) examples.
This patch adds examples to show how to use DB API.

The number of new examples are three.
 - insert
 - find
 - findOne

To run insert example, send POST to /plugins/insert/{key}/{value}.
This example insert key and value to example.keyValue collection.

To run find example, send POST to /plugins/find/{key1}[/{keyN}]
This example find all documents that the key is one of keyN.

To run findOne example, send POST to /plugins/findOne/{key1}[/{keyN}]
This example find a document that the key is one of keyN.