Adjust relations between Scala, JavaScript and BSON types.
Currently scala types, JavaScript types and BSON types are not
consistent. This patch adjusts their relations.

There are three kinds of type.
 - JavaScript type: used by rhino engine
 - Scala type: helper type to make scala code
 - BSON type: used by database

And a type table is
 JavaScriptType  | ScalaType            | BSONType
-----------------+----------------------+----------------
 NativeObject    | Map[String, AnyRef]  | BSONDocument
 NativeArray     | Seq[AnyRef]          | BSONArray
 NativeNumber    | Int                  | BSONInteger
                 | Long                 | BSONLong
                 | Double               | BSONDouble
 NativeBoolean   | Boolean              | BSONBoolean
 NativeDate      | Date                 | BSONDateTime
 NativeUndefined | Undefined            | BSONUndefined
 NativeFunction  | String               | BSONJavaScript

JavaScript types can be converted to Scala types with Field and Scala
types can be converted to JavaScript types with Context and Scope.
Scala types are convertible to BSON types and vice versa.
But JavaScript types and BSON types cannot be converted to each other
directly.

ScriptableDocument contains data as BSON Type, and converts to
JAvaScript Type when engine requests it.

Scala types seem to be useless because it's not used by engine nor
database, but code becomes complicate without this types because the
Rhino engine automatically converts primitive types with toSource method
of Native*.