Get rid of JavaScript timer functions.
setTimer and setInterval are rarely used. Actually, these functions should not
be used in handler functions because scheduling multiple timers at the same
time has a negative impact on performance. Because the philosophy of Beyond is
to systematically discourage bad programming idioms, I think it's better to get
rid of JavaScript timer functions entirely.