Make ScriptableQuery handle ScriptableObjectId.
Currently, ScriptableQuery cannot handle ScriptableObjectId because it uses
AnyRefBSONHandler directly. We need to make AnyRefBSONHandler handle
ScriptableObjectId or make ScriptableQuery change ScriptableObjectId to
ObjectId before using AnyRefBSONHandler.

I chose the second option because AnyRefBSONHandler is used by
ScriptableDocument with convertJavaScriptToScalaWithField method.
ScriptableObjectId is not passed to AnyRefBSONHandler in this case.