Make ScriptableQuery handle ScriptableObjectId.
Currently, ScriptableQuery cannot handle ScriptableObjectId, because it
uses AnyRefBSONHandler directly. We need to make AnyRefBSONHandler
handle ScriptableObjectId or ScriptableQuery changes ScriptableObjectId
to ObjectId before using AnyRefBSONHandler.

I decided the second choice because AnyRefBSONHandler is used by
ScriptableDocument with convertJavaScriptToScalaWithField method. And
ScriptableObjectId is not passed to AnyRefBSONHandler in this case.