Modify response.js to use ResponseInternal directly.
The logic for the functionality was separated into a JavaScript and
Scala files. In this case, if we wanted to change the logic, we would
have to change both of them. This patch merged the logic in the
JavaScript file into the Scala one.

Also a test case for ScriptableResponse has been added. Some properties
have also been added to test the functionalities.