Fix regression errors by the previous ScriptableResponse patch.
This change should have been included in the previous commit. This
commit fixes a regression problem the previous commit causes.

First of all, this fixed GamePlugin to handle ScriptableResponse
directly. It originally handled it as a Scriptable object and found
the `_response` property from it.

Also, this removed the fallback type of stringify, because it could be a
JavaScript Number or null, both of which are not Scriptable.