Seperate module source provider for library and plugin code.
Currently, both library and plugin uses UrlModuleSourceProvider. But
assets are packaged to jar on deployment and UrlModuleSourceProvider
cannot read source from jar. So this patch makes Beyond uses
AssetsModuleSourceProvider for library.
But plugin code still should read with UrlModuleSourceProvider.

To implement it this patch makes BeyondGlobal takes ModuleSourceProvider
as an argument. This provider will be used for library code.