Add `beyond.mongodb.type` option for MongoDB.
Currently Beyond launches a standalone MongoDB server on localhost. This
patch renames the current MongoDBLauncher to MongoDBStandaloneLauncher
and makes Beyond use it only when `beyond.mongodb.type` is "standalone".

There could also be "config", "routing" and "shard" for the type to
support MongoDB sharding, but they are not yet supported or handled as
the sharding itself hasn't been implemented.