Change optional option to nullable.
Currently, we don't put the value for optional, but there is need to save
something for the value of array. (e.g. `type: 'array', elementType: {
type: 'int', optional: true }`)

We determined to put null for optional value.
And this patch also changes the option name to nullable, because optional can
be mistaken. But keeping optional option for a while to support backward
compatibility.