Support clustered ZooKeeper server.
Currently, beyond launches ZooKeeper as standalone mode.
This patch makes beyond launch ZooKeeper as clustered mode when there
are server lists in the zookeeper config file and as standalone mode
when there is no server lists in the zookeeper config file. This
behavior is implemented on QuorumPeerMain. So this patch just modifies to use
QuorumPeerMain instead of ZooKeeperServerMain.

You have to add `dataDir`/myid file to use clustered mode. It's
ZooKeeper's restriction. Read
http://zookeeper.apache.org/doc/r3.4.6/zookeeperStarted.html#sc_RunningReplicatedZooKeeper
for more information.