Implement Configuration.
Currently, we prepare to use spray instead of the play framework. See this link
(https://github.com/SollmoStudio/beyond/issues/210) for more information.

This is the preparation for replacing.

This is copied from play framework. (https://github.com/playframework/playframework)
But not fully copied. This patch just copied only the current beyond
framework needed.

And add a implicit converter to convert play Configuration to beyond Configuration.